Move pause and game speed to handle_input
